some code cleanup
git-svn-id: https://yabause.svn.sourceforge.net/svnroot/yabause/trunk@151 d7be59e3-070e-4b9a-85a1-4646befae34d

+++ yabause/src/vdp1.hh
@@ -53,7 +53,7 @@ class Vdp2;
 class Vdp2ColorRam;
 
 typedef struct {
-	int vdp1_loc;
+	unsigned long vdp1_loc;
 	bool dirty;
 	GLuint txr;
 } vdp1Sprite;
